requirements.txt
aiogram==3.1.0
asyncpg==0.27.0
flyerapi==0.1.7
aiohttp==3.8.5
python-dotenv==1.0.0